Move grpc header logic under the workloadapi package
