+++ src/pyspiffe/workloadapi/grpc/__init__.py
@@ -1,0 +1,3 @@
+"""
+grpc Module. Contains logic related to the grpc connections.
+"""
