+++ src/pyspiffe/workloadapi/grpc/generic_client_interceptor.py
@@ -3,38 +3,44 @@
 import grpc
 
 
-class _GenericClientInterceptor(grpc.UnaryUnaryClientInterceptor,
-                                grpc.UnaryStreamClientInterceptor,
-                                grpc.StreamUnaryClientInterceptor,
-                                grpc.StreamStreamClientInterceptor):
-
+class _GenericClientInterceptor(
+    grpc.UnaryUnaryClientInterceptor,
+    grpc.UnaryStreamClientInterceptor,
+    grpc.StreamUnaryClientInterceptor,
+    grpc.StreamStreamClientInterceptor,
+):
     def __init__(self, interceptor_function):
         self._fn = interceptor_function
 
     def intercept_unary_unary(self, continuation, client_call_details, request):
         new_details, new_request_iterator, postprocess = self._fn(
-            client_call_details, iter((request,)), False, False)
+            client_call_details, iter((request,)), False, False
+        )
         response = continuation(new_details, next(new_request_iterator))
         return postprocess(response) if postprocess else response
 
-    def intercept_unary_stream(self, continuation, client_call_details,
-                               request):
+    def intercept_unary_stream(self, continuation, client_call_details, request):
         new_details, new_request_iterator, postprocess = self._fn(
-            client_call_details, iter((request,)), False, True)
+            client_call_details, iter((request,)), False, True
+        )
         response_it = continuation(new_details, next(new_request_iterator))
         return postprocess(response_it) if postprocess else response_it
 
-    def intercept_stream_unary(self, continuation, client_call_details,
-                               request_iterator):
+    def intercept_stream_unary(
+        self, continuation, client_call_details, request_iterator
+    ):
         new_details, new_request_iterator, postprocess = self._fn(
-            client_call_details, request_iterator, True, False)
+            client_call_details, request_iterator, True, False
+        )
         response = continuation(new_details, new_request_iterator)
         return postprocess(response) if postprocess else response
 
-    def intercept_stream_stream(self, continuation, client_call_details,
-                                request_iterator):
+    def intercept_stream_stream(
+        self, continuation, client_call_details, request_iterator
+    ):
         new_details, new_request_iterator, postprocess = self._fn(
-            client_call_details, request_iterator, True, True)
+            client_call_details, request_iterator, True, True
+        )
         response_it = continuation(new_details, new_request_iterator)
         return postprocess(response_it) if postprocess else response_it
