+++ src/pyspiffe/workloadapi/grpc/header_manipulator_client_interceptor.py
@@ -1,0 +1,40 @@
+"""Interceptor that adds headers to outgoing requests."""
+
+import collections
+
+import grpc
+
+from pyspiffe.workloadapi.grpc import generic_client_interceptor
+
+
+class _ClientCallDetails(
+    collections.namedtuple(
+        '_ClientCallDetails', ('method', 'timeout', 'metadata', 'credentials')
+    ),
+    grpc.ClientCallDetails,
+):
+    pass
+
+
+def header_adder_interceptor(header, value):
+    def intercept_call(
+        client_call_details, request_iterator, request_streaming, response_streaming
+    ):
+        metadata = []
+        if client_call_details.metadata is not None:
+            metadata = list(client_call_details.metadata)
+        metadata.append(
+            (
+                header,
+                value,
+            )
+        )
+        client_call_details = _ClientCallDetails(
+            client_call_details.method,
+            client_call_details.timeout,
+            metadata,
+            client_call_details.credentials,
+        )
+        return client_call_details, request_iterator, None
+
+    return generic_client_interceptor.create(intercept_call)
